+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - 🧨 Jena is upgraded to 4.1.0
   - Jena renamed `RDFReader/RDFWriter` to `RDFReaderI/RDFWriterI`
   - LyoStore: Ordering resources by their subject IDs when doing a query to store. This ordering can be disabled with a call to `OSLC4JUtils.setLyoStorePagingUnsafe(true)`
+  - LyoStore: Providing an OSLC4JUtils property (set/hasLyoStorePagingPreciseLimit) to allow the application to indicate whether query paging should return an exact number of elements in a paged query, or not.
 
 
 ### Deprecated